Creating insight package.
