+++ src/com/google/bamboo/soy/insight/completion/SoyCompletionContributor.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.bamboo.soy.completion;
+package com.google.bamboo.soy.insight.completion;
 
 import static com.intellij.patterns.PlatformPatterns.psiElement;
