+++ src/com/google/bamboo/soy/insight/documentation/SoyDocumentationProvider.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.bamboo.soy.documentation;
+package com.google.bamboo.soy.insight.documentation;
 
 import com.google.bamboo.soy.parser.SoyAtInjectSingle;
 import com.google.bamboo.soy.parser.SoyAtParamSingle;
