+++ src/com/google/bamboo/soy/insight/highlight/SoySyntaxHighlighter.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.bamboo.soy.highlight;
+package com.google.bamboo.soy.insight.highlight;
 
 import static com.intellij.openapi.editor.colors.TextAttributesKey.createTextAttributesKey;
