+++ src/com/google/bamboo/soy/insight/structure/SoyStructureViewModel.kt
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.bamboo.soy.structure
+package com.google.bamboo.soy.insight.structure
 
 import com.google.bamboo.soy.file.SoyFile
 import com.intellij.ide.structureView.StructureViewModel
