+++ src/com/google/bamboo/soy/insight/typedhandlers/ClosingTagHandler.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.bamboo.soy.typedhandlers;
+package com.google.bamboo.soy.insight.typedhandlers;
 
 import com.google.bamboo.soy.BracedTagUtils;
 import com.google.bamboo.soy.parser.SoyParamListElement;
