+++ src/com/google/bamboo/soy/insight/typedhandlers/EnterHandler.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.bamboo.soy.typedhandlers;
+package com.google.bamboo.soy.insight.typedhandlers;
 
 import com.google.bamboo.soy.file.SoyFile;
 import com.google.bamboo.soy.file.SoyFileType;
